Updating Based on LHs recommendations
tried to remedy the weird year where everything took a dramatic turn. Looks like it's an outlier problem, so as a quick fix I just used the medians instead of the means to aggregate.

Also updated the paths to sync up with the data files in the repository. Altered the data cleaning code accordingly.

Also reanalyzed everything with the new median aggregations. Made and saved a network for JR and KC with short and long VRs.